rename electron and lattice modules by prepending underscore
