+++ tests/covalent_dispatcher_tests/_core/dispatcher_test.py
@@ -30,7 +30,7 @@ import pytest
 
 import covalent as ct
 from covalent._results_manager import Result
-from covalent._workflow.lattice import Lattice
+from covalent._workflow._lattice import Lattice
 from covalent_dispatcher._core.dispatcher import (
     _get_abstract_task_inputs,
     _get_initial_tasks_and_deps,
