+++ tests/functional_tests/serialization_test.py
@@ -23,7 +23,7 @@ import covalent
 import covalent as cova
 from covalent import electron as etron
 from covalent._shared_files.utils import get_serialized_function_str
-from covalent._workflow.lattice import Lattice
+from covalent._workflow._lattice import Lattice
 
 executor = covalent.executor._executor_manager.list_executors(print_names=False)[0]
